chore: flatten publishing workflow runner path
