+++ scripts/publishingWorkflowRunner.ts
@@ -1,8 +1,10 @@
 import fs from 'fs';
 import path from 'path';
 
-import { readContentFile, type ParsedContent } from '../content/front_matter';
-import { HttpError, PlatformValue, loadPostMap, savePostMap } from '../utils';
+import { readContentFile, type ParsedContent } from './utils/frontMatter';
+import { HttpError } from './utils/http';
+import { PlatformValue } from './utils/platform';
+import { loadPostMap, savePostMap } from './utils/postMap';
 
 export type BaseFrontMatter = {
   title: string;
