+++ scripts/qiita.ts
@@ -1,8 +1,9 @@
 import fetch from 'node-fetch';
 
-import { createHttpError, ensureArrayOfStrings, wantsPlatform } from './utils';
-import { PublishingAdapter, runPublishingWorkflow, ValidationResult } from './workflows/base';
-import type { QiitaFrontMatter } from './content/front_matter';
+import { createHttpError } from './utils/http';
+import { ensureArrayOfStrings, type QiitaFrontMatter } from './utils/frontMatter';
+import { wantsPlatform } from './utils/platform';
+import { PublishingAdapter, runPublishingWorkflow, ValidationResult } from './publishingWorkflowRunner';
 
 type QiitaPostMapEntry = {
   id: string;
